+++ CHANGELOG.md
@@ -81,6 +81,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added function `SKY_bip44_Coin_Account`
 - Added function `SKY_bip44_Account_External`
 - Added function `SKY_bip44_Account_Change`
+- Added function `SKY_bip44_Account_String`
+- Added function `SKY_bip44_Account_GetPrivateKey`
 
 ### Removed